+++ data/settings.js
@@ -135,9 +135,14 @@
         VIP: 7000,
         VLC: 4000,
         VPC: 6500,
+        VPI: 6000,
         VCG: 4500,
         VBS: 4500,
-        VLSM: 10000
+        VLSM: 10000,
+        CCGC: 13000,
+        CCFL: 9000,
+        CCRL: 8000,
+        CCFS: 16000
       },
       casernes: {
         ORL_NOR: 20000,
@@ -172,15 +177,20 @@
         VIP: 10000,
         VLC: 6500,
         VPC: 9000,
+        VPI: 10000,
         VCG: 7000,
         VBS: 7000,
-        VLSM: 15500
+        VLSM: 15500,
+        CCGC: 23000,
+        CCFL: 14500,
+        CCRL: 13500,
+        CCFS: 28000
       }
     }
   }
 };
  const APP_META = {
   name: "CTA-Manager Lite",
-  version: "v0.14.1"
+  version: "v0.14.2"
 };
 window.APP_META = APP_META;